sched/cass: Don't pack tasks with uclamp boosts below minimum CPU capacity To save energy, CASS may prefer non-idle CPUs for uclamp-boosted tasks in order to pack them onto a single performance domain rather than spreading them across multiple performance domains. This way, it is more likely for only one performance domain to be boosted a higher P-state when there is more than one uclamp-boosted task running.
However, when a task has a uclamp boost value that is below a CPU's minimum
capacity, it is nearly the same thing as not having a uclamp boost at all.

In spite of that, CASS may still prefer non-idle CPUs for tasks with bogus
uclamp boost values. This is not only worse for latency, but also energy
efficiency since the load on the CPU is spread less evenly as a result.

Therefore, don't pack tasks with uclamp boosts below a CPU's minimum
configured capacity, since such tasks do not force the CPU to run at a
higher P-state.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>